Extract the dummy user code from the tests.
They could be use somewhere else.
